+++ src/platforms/default/Default.ts
@@ -154,7 +154,7 @@ Platform.load({
     server: true,
     cache: new CacheWithMap(),
     fileCache: new FileCache(),
-    fetcher: fetch,
+    fetcher: (url, options) => fetch(url, options),
     poToken: () => Promise.resolve({ poToken: '', visitorData: '' }),
     default: {
         options: {